refactor: 패키지를 모두 lotto로 이동
