+++ src/main/java/lotto/domain/lotto/DefaultRandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.Collections;
 import java.util.List;
