+++ src/main/java/lotto/domain/lotto/QuickPickStrategy.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.List;
 import java.util.Objects;
