+++ src/test/java/lotto/domain/lotto/Fixture.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 public class Fixture {
     public static NumberPickStrategy pick123456() {
