+++ src/test/java/lotto/domain/lotto/Pick123456Strategy.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.Arrays;
 import java.util.List;
