+++ src/test/java/lotto/domain/lotto/QuickPickStrategyTest.java
@@ -1,10 +1,10 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 import java.util.Arrays;
 import java.util.List;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -30,6 +30,6 @@ class QuickPickStrategyTest {
         );
 
         final List<Lotto> actual = strategy.pickNumbers(quantity);
-        assertThat(actual).containsExactlyElementsOf(expected);
+        Assertions.assertThat(actual).containsExactlyElementsOf(expected);
     }
 }
